VendorOrderTracker/Models/Order.cs
using System.Collections.Generic;
using System;

namespace VendorOrderTracker.Models
{
  public class Order
  {
    private static List<Order> _instances = new List<Order> { };

    public string Title { get; set; }
    public string Description { get; set; }
    public int Price { get; set; }
    public DateTime Date { get; set; }

    public Order(string title, string description, int price, DateTime date)
    {
      Title = title;
      Description = description;
      Price = price;
      Date = date;
    }
    public static List<Order> GetAll()
    {
      return _instances;
    }
  }
}